+++ branches/2.0/doubango/plugins/pluginDirectShow/plugin_video_dshow_producer.cxx
@@ -144,7 +144,7 @@ static int plugin_video_dshow_producer_start(tmedia_producer_t* self)
 	producer->grabber->setCaptureDevice("Null");
 
 	// set parameters
-	producer->grabber->setCaptureParameters(TMEDIA_PRODUCER(producer)->video.width, TMEDIA_PRODUCER(producer)->video.height, TMEDIA_PRODUCER(producer)->video.fps);
+	producer->grabber->setCaptureParameters((int)TMEDIA_PRODUCER(producer)->video.width, (int)TMEDIA_PRODUCER(producer)->video.height, TMEDIA_PRODUCER(producer)->video.fps);
 
 	// set callback function
 	producer->grabber->setCallback(plugin_video_dshow_plugin_cb, producer);
@@ -154,7 +154,7 @@ static int plugin_video_dshow_producer_start(tmedia_producer_t* self)
 		if(producer->previewHwnd){
 			producer->grabber->preview->attach(producer->previewHwnd);
 		}
-		producer->grabber->preview->setSize(TMEDIA_PRODUCER(producer)->video.width, TMEDIA_PRODUCER(producer)->video.height);
+		producer->grabber->preview->setSize((int)TMEDIA_PRODUCER(producer)->video.width, (int)TMEDIA_PRODUCER(producer)->video.height);
 	}
 	
 	// start grabber
